requirements.txt
google-genai
websockets